+++ core/src/main/java/org/kohsuke/stapler/CrumbIssuer.java
@@ -27,7 +27,7 @@ public abstract class CrumbIssuer {
      * Sends the crumb value in plain text, enabling retrieval through XmlHttpRequest.
      */
     public HttpResponse doCrumb() {
-        return HttpResponses.plainText(issueCrumb());
+        return HttpResponses.text(issueCrumb());
     }
 
     /**